+++ collector-bitmovin-player/src/main/java/com/bitmovin/analytics/bitmovin/player/BitmovinUtil.java
@@ -4,9 +4,9 @@ import com.bitmovin.analytics.utils.Util;
 import com.bitmovin.player.api.Player;
 import java.lang.reflect.Field;
 
-public class BitmovinUtil {
+class BitmovinUtil {
 
-    public static String getPlayerVersion() {
+    static String getPlayerVersion() {
         try {
             Field versionField = com.bitmovin.player.BuildConfig.class.getField("VERSION_NAME");
             return (String) versionField.get(null);
@@ -16,7 +16,7 @@ public class BitmovinUtil {
         return "unknown";
     }
 
-    public static long getCurrentTimeInMs(Player player) {
+    static long getCurrentTimeInMs(Player player) {
         return Util.secondsToMillis(player.getCurrentTime());
     }
 }